requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.0
pydantic-settings==2.6.1
pytest==8.3.3
httpx==0.27.2
